requirements.txt
mysql-connector-python==8.2.0
graphviz==0.20.1
